Converted to middleman 4 - tests pass
