+++ lib/middleman-cloudfront/extension.rb
@@ -1,36 +1,37 @@
 require 'middleman-core'
+require 'middleman-cloudfront/commands/invalidate'
 
 module Middleman
   module CloudFront
-    class Options < Struct.new(:access_key_id, :secret_access_key, :distribution_id, :filter, :after_build); end
+    class Extension < Middleman::Extension
+      # @param [Symbol] key The name of the option
+      # @param [Object] default The default value for the option
+      # @param [String] description A human-readable description of what the option does
+      option :access_key_id, nil, 'Access key id'
+      option :secret_access_key, nil, 'Secret access key'
+      option :distribution_id, nil, 'Distribution id'
+      option :filter, /.*/, 'Filter files to be invalidated'
+      option :after_build, false, 'Invalidate after build'
 
-    class << self
-      def options
-        @@cloudfront_options
+      def initialize(app, options_hash={}, &block)
+        super
       end
 
-      def registered(app, options_hash = {}, &block)
-        @@cloudfront_options = Options.new(options_hash)
-        yield @@cloudfront_options if block_given?
+      def after_build
+        ::Middleman::Cli::CloudFront.new.invalidate(options) if options.after_build
+      end
 
-        app.after_build do
-          ::Middleman::Cli::CloudFront.new.invalidate(@@cloudfront_options) if @@cloudfront_options.after_build
+      # expose_to_application :cloudfront_options => :options
+      helpers do
+        # def cloudfront_options
+        #   options
+        # end
+
+        def invalidate(files = nil)
+          # extension_options = app.extensions[:cloudfront][options[:cloudfront]]
+          ::Middleman::Cli::CloudFront.new.invalidate(options, files)
         end
-
-        app.send :include, Helpers
-      end
-      alias :included :registered
-    end
-
-    module Helpers
-      def cloudfront_options
-        ::Middleman::CloudFront.options
-      end
-
-      def invalidate(files = nil)
-        ::Middleman::Cli::CloudFront.new.invalidate(cloudfront_options, files)
       end
     end
-
   end
 end
